requirements.txt



SQLAlchemy>=2.0,<3
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]